backend/requirements.txt
streamlit 
moviepy 
assemblyai 
google-genai 
pandas 
python-dotenv
ollama
pydantic
pymongo[srv]
plotly
altair
uvicorn
fastapi
python-multipart
